--- Tests/PlayMode/RemoteTrackEnabledRegressionTests.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b2303181ca4ca4c6888a14b77b687aae
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
